Moved CustomAllocator code from vsgallocator example to core vsg::Allocator implementation.
